Retry setting up connections after SIGHUP a few times
It might take a bit for configuration values etc. to be updated everywhere, so
just give it a few seconds.

Use a custom exception class to stop the world after reloading settings during
SIGHUP and just start the mainloop again.

This also moves complicated logic out of the signal handler into its own
reload_settings function that's called in the signal handler and at the
beginning of mainloop(). This allows moving the retry-on-pg-error logic into
run.py