Reduce the log output of the bot
This changes all output during normal operation to log level DEBUG. This reduces
the per-entity messages to zero. This also changes errors and warnings to log
levels ERROR/WARN.

An additional message is added when switching from one entity type to another
with the amount of entities of that type that are going to be processed, to give
a rough estimate of how long things should take (hours? days? weeks?).