Start adding work Wikidata pages again
Work Wikipedia pages have issues - they're often
for both RGs and works, especially ones for singles, and as such
the bot ends up linking RGs to Wikidata work pages.
Mineo stopped the bot from adding IDs to works entirely because of that
but that seems like overkill - there's plenty of classical works
for example which should be linked. As such, this allows linking,
but only if MB stores a Wikidata link directly.
That should hopefully stop most of the wrong links
and it's the approach we recently took for RG links as well.