Set P1810 (Named as) on Wikidata claims
I was asked to add this for genres by an editor who is doing it by hand
but I cannot actually see a reason why we would not want
to do this for all cases.
It also makes it a lot easier to spot problems later on
where the wrong link was created.