Add qualifier and sources with claim
This is apparently possible now (it used to be that you needed to do it
after the claim was added to the page).
This way seems a lot more clean.
I've seen some issues with claims being added twice,
maybe this will help with that? But it seems good either way.